refactor: extract XLPivotFieldBase to deduplicate pivot field implementations
Extract shared fluent setters and property delegates from XLPivotTableAxisField
and XLPivotTablePageField into a common XLPivotFieldBase abstract base class.
Also fixes SetCollapsed bug in AxisField that ignored the value parameter.
