+++ XLibur/Excel/PivotTables/Areas/XLPivotTablePageField.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -8,176 +8,25 @@ namespace XLibur.Excel;
 /// Fluent API for filter fields of a <see cref="XLPivotTable"/>. This class shouldn't contain any
 /// state, only logic to change state per API.
 /// </summary>
-internal sealed class XLPivotTablePageField : IXLPivotField
+internal sealed class XLPivotTablePageField : XLPivotFieldBase
 {
-    private readonly XLPivotTable _pivotTable;
     private readonly XLPivotPageField _filterField;
 
     internal XLPivotTablePageField(XLPivotTable pivotTable, XLPivotPageField filterField)
+        : base(pivotTable)
     {
-        _pivotTable = pivotTable;
         _filterField = filterField;
     }
 
-    public string SourceName => _pivotTable.PivotCache.FieldNames[_filterField.Field];
+    public override string SourceName => PivotTable.PivotCache.FieldNames[_filterField.Field];
 
-    public string CustomName
+    public override string CustomName
     {
         get => GetField().Name!;
         set => GetField().Name = value;
     }
 
-    public string SubtotalCaption
-    {
-        get => GetField().SubtotalCaption;
-        set => GetField().SubtotalCaption = value;
-    }
-
-    public IReadOnlyCollection<XLSubtotalFunction> Subtotals => GetField().Subtotals;
-
-    public bool IncludeNewItemsInFilter
-    {
-        get => GetField().IncludeNewItemsInFilter;
-        set => GetField().IncludeNewItemsInFilter = value;
-    }
-
-    public bool Outline
-    {
-        get => GetField().Outline;
-        set => GetField().Outline = value;
-    }
-
-    public bool Compact
-    {
-        get => GetField().Compact;
-        set => GetField().Compact = value;
-    }
-
-    public bool? SubtotalsAtTop
-    {
-        get => GetField().SubtotalTop;
-        set => GetField().SubtotalTop = value ?? true;
-    }
-
-    public bool RepeatItemLabels
-    {
-        get => GetField().RepeatItemLabels;
-        set => GetField().RepeatItemLabels = value;
-    }
-
-    public bool InsertBlankLines
-    {
-        get => GetField().InsertBlankRow;
-        set => GetField().InsertBlankRow = value;
-    }
-
-    public bool ShowBlankItems
-    {
-        get => GetField().ShowAll;
-        set => GetField().ShowAll = value;
-    }
-
-    public bool InsertPageBreaks
-    {
-        get => GetField().InsertPageBreak;
-        set => GetField().InsertPageBreak = value;
-    }
-
-    public bool Collapsed
-    {
-        get => GetField().Collapsed;
-        set => GetField().Collapsed = value;
-    }
-
-    public XLPivotSortType SortType
-    {
-        get => GetField().SortType;
-        set => GetField().SortType = value;
-    }
-
-    public IXLPivotField SetCustomName(string value)
-    {
-        CustomName = value;
-        return this;
-    }
-
-    public IXLPivotField SetSubtotalCaption(string value)
-    {
-        SubtotalCaption = value;
-        return this;
-    }
-
-    public IXLPivotField SetSubtotal(XLSubtotalFunction function, bool enabled)
-    {
-        if (enabled)
-            GetField().AddSubtotal(function);
-        else
-            GetField().RemoveSubtotal(function);
-
-        return this;
-    }
-
-    public IXLPivotField AddSubtotal(XLSubtotalFunction value)
-    {
-        GetField().AddSubtotal(value);
-        return this;
-    }
-
-    public IXLPivotField SetIncludeNewItemsInFilter(bool value = true)
-    {
-        IncludeNewItemsInFilter = value;
-        return this;
-    }
-
-    public IXLPivotField SetLayout(XLPivotLayout value)
-    {
-        GetField().SetLayout(value);
-        return this;
-    }
-
-    public IXLPivotField SetSubtotalsAtTop(bool value = true)
-    {
-        SubtotalsAtTop = value;
-        return this;
-    }
-
-    public IXLPivotField SetRepeatItemLabels(bool value = true)
-    {
-        RepeatItemLabels = value;
-        return this;
-    }
-
-    public IXLPivotField SetInsertBlankLines(bool value = true)
-    {
-        InsertBlankLines = value;
-        return this;
-    }
-
-    public IXLPivotField SetShowBlankItems(bool value = true)
-    {
-        ShowBlankItems = value;
-        return this;
-    }
-
-    public IXLPivotField SetInsertPageBreaks(bool value = true)
-    {
-        InsertPageBreaks = value;
-        return this;
-    }
-
-    public IXLPivotField SetCollapsed(bool value = true)
-    {
-        Collapsed = value;
-        return this;
-    }
-
-    public IXLPivotField SetSort(XLPivotSortType value)
-    {
-        SortType = value;
-        return this;
-    }
-
-    public IReadOnlyList<XLCellValue> SelectedValues
+    public override IReadOnlyList<XLCellValue> SelectedValues
     {
         get
         {
@@ -194,7 +43,7 @@ internal sealed class XLPivotTablePageField : IXLPivotField
         }
     }
 
-    public IXLPivotField AddSelectedValue(XLCellValue value)
+    public override IXLPivotField AddSelectedValue(XLCellValue value)
     {
         // Try to keep the original behavior of XLibur - it always allows multiple selected items for added values.
         // But it's complete kludge with no sane semantic that will be nuked ASAP.
@@ -232,7 +81,7 @@ internal sealed class XLPivotTablePageField : IXLPivotField
         }
     }
 
-    public IXLPivotField AddSelectedValues(IEnumerable<XLCellValue> values)
+    public override IXLPivotField AddSelectedValues(IEnumerable<XLCellValue> values)
     {
         foreach (var value in values)
             AddSelectedValue(value);
@@ -240,14 +89,14 @@ internal sealed class XLPivotTablePageField : IXLPivotField
         return this;
     }
 
-    public IXLPivotFieldStyleFormats StyleFormats => throw new NotImplementedException("Styles for filter fields are not yet implemented.");
-    public bool IsOnRowAxis => false;
-    public bool IsOnColumnAxis => false;
-    public bool IsInFilterList => true;
-    public int Offset => _filterField.Field;
+    public override IXLPivotFieldStyleFormats StyleFormats => throw new NotImplementedException("Styles for filter fields are not yet implemented.");
+    public override bool IsOnRowAxis => false;
+    public override bool IsOnColumnAxis => false;
+    public override bool IsInFilterList => true;
+    public override int Offset => _filterField.Field;
 
-    private XLPivotTableField GetField()
+    private protected override XLPivotTableField GetField()
     {
-        return _pivotTable.PivotFields[_filterField.Field];
+        return PivotTable.PivotFields[_filterField.Field];
     }
 }
